[Enhancement] Changes variant count types from int32_t to size_t (#1498)
This is a repeat of
https://github.com/celonis/celostar-starrocks/pull/1494 due to a default
branch switch.

## Why I'm doing:

While I tested process mining operators, I noticed we used 32bit counts
for variants.
In the walgreen e2e test data model, out of 1.3B rows in EKPO, one
variant had 0.85B rows.
As we have larger data models, 2.1B limit is not safe.

## What I'm doing:

Updates count types of VariantHashMap and VariantCount from int32_t to
size_t.
In other parts of the code using them like serialization, we already use
size_t. So the change turns out to be simple and the impact is expected
to be non.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [X] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

+++ be/src/exprs/celonis/agg/variant.h
@@ -38,13 +38,13 @@ struct HashOnVariant {
 };
 
 using SliceHashMap = phmap::flat_hash_map<SliceWithHash, int32_t, HashOnSliceWithHash, EqualOnSliceWithHash>;
-using VariantHashMap = phmap::flat_hash_map<Variant, int32_t, HashOnVariant, EqualOnVariant>;
+using VariantHashMap = phmap::flat_hash_map<Variant, size_t, HashOnVariant, EqualOnVariant>;
 
 struct VariantCount {
-    VariantCount(std::vector<int32_t> v, int32_t c) : variant(std::move(v)), count(c) {}
+    VariantCount(std::vector<int32_t> v, size_t c) : variant(std::move(v)), count(c) {}
 
     std::vector<int32_t> variant;
-    int32_t count;
+    size_t count;
 
     bool operator<(const VariantCount& other) const { return variant < other.variant; }
 };
